Unifies QuadTree and QuadNode
Previously a QuadTree stored a pointer to a QuadNode root, which stored
its children. However, each QuadNode would have been a valid root, which
made the architecture overly complex. Additionally, a great deal of
fucntions in the QuadTree simply just called a node function on the
root. This cuts out the middleman struct and function calls.